Enable speedLimiter and netInfo
