+++ core/protocol.go
@@ -5,9 +5,9 @@ import (
 	"encoding/binary"
 	"encoding/json"
 	"hash/crc64"
-	"math/big"
-	"net"
 	"time"
+
+	"openp2p/util"
 )
 
 const OpenP2PVersion = "3.12.0"
@@ -326,14 +326,14 @@ type RelayHeartbeat struct {
 }
 
 type ReportBasic struct {
-	OS              string  `json:"os,omitempty"`
-	Mac             string  `json:"mac,omitempty"`
-	LanIP           string  `json:"lanIP,omitempty"`
-	HasIPv4         int     `json:"hasIPv4,omitempty"`
-	IPv6            string  `json:"IPv6,omitempty"`
-	HasUPNPorNATPMP int     `json:"hasUPNPorNATPMP,omitempty"`
-	Version         string  `json:"version,omitempty"`
-	NetInfo         NetInfo `json:"netInfo,omitempty"`
+	OS              string       `json:"os,omitempty"`
+	Mac             string       `json:"mac,omitempty"`
+	LanIP           string       `json:"lanIP,omitempty"`
+	HasIPv4         int          `json:"hasIPv4,omitempty"`
+	IPv6            string       `json:"IPv6,omitempty"`
+	HasUPNPorNATPMP int          `json:"hasUPNPorNATPMP,omitempty"`
+	Version         string       `json:"version,omitempty"`
+	NetInfo         util.NetInfo `json:"netInfo,omitempty"`
 }
 
 type ReportConnect struct {
@@ -398,25 +398,6 @@ type UpdateInfo struct {
 	Error       int    `json:"error,omitempty"`
 	ErrorDetail string `json:"errorDetail,omitempty"`
 	Url         string `json:"url,omitempty"`
-}
-
-type NetInfo struct {
-	IP         net.IP   `json:"ip"`
-	IPDecimal  *big.Int `json:"ip_decimal"`
-	Country    string   `json:"country,omitempty"`
-	CountryISO string   `json:"country_iso,omitempty"`
-	CountryEU  *bool    `json:"country_eu,omitempty"`
-	RegionName string   `json:"region_name,omitempty"`
-	RegionCode string   `json:"region_code,omitempty"`
-	MetroCode  uint     `json:"metro_code,omitempty"`
-	PostalCode string   `json:"zip_code,omitempty"`
-	City       string   `json:"city,omitempty"`
-	Latitude   float64  `json:"latitude,omitempty"`
-	Longitude  float64  `json:"longitude,omitempty"`
-	Timezone   string   `json:"time_zone,omitempty"`
-	ASN        string   `json:"asn,omitempty"`
-	ASNOrg     string   `json:"asn_org,omitempty"`
-	Hostname   string   `json:"hostname,omitempty"`
 }
 
 type ProfileInfo struct {
